+++ apps/frontend/src/assets/index.ts
@@ -1,7 +1,7 @@
 // Images
-export { default as starLogo } from './images/star.svg';
-export { default as starLogoWhite } from './images/star-white.svg';
-export { default as githubLogo } from './images/github-mark.svg';
-export { default as githubLogoWhite } from './images/github-mark-white.svg';
-export { default as forkLogo } from './images/fork.svg';
-export { default as forkLogoWhite } from './images/fork-white.svg';
+export { default as starLogo } from './images/star.svg'
+export { default as starLogoWhite } from './images/star-white.svg'
+export { default as githubLogo } from './images/github-mark.svg'
+export { default as githubLogoWhite } from './images/github-mark-white.svg'
+export { default as forkLogo } from './images/fork.svg'
+export { default as forkLogoWhite } from './images/fork-white.svg'